admin schedule

+++ app/Http/Controllers/Admin/ProductController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\OrderItem;
 use App\Product;
+use App\Schedule;
 use Illuminate\Http\Request;
 use Yajra\DataTables\DataTables;
 
@@ -44,20 +46,23 @@ class ProductController extends Controller
             $product->price = $request->input('price');
             $product->rrp_price = $request->input('rrp_price');
             $product->available = $request->input('available');
-            $product->booking = $request->input('booking') ? 1:0;
             $product->schedule_id = $request->input('schedule_id');
+            $product->booking = $product->schedule_id ? 1:0;
 
             $product->save();
             return view('admin.productlist');
         }
 
-        return view('admin.productadd', compact('product'));
+        $schedules = Schedule::where('available', 1)->get();
+
+        return view('admin.productadd', compact('product', 'schedules'));
     }
 
     public function edit(Request $request, $id) {
         $product = Product::find($id);
         if($product) {
-            return view('admin.productadd', compact('product'));
+            $schedules = Schedule::where('available', 1)->get();
+            return view('admin.productadd', compact('product','schedules'));
         }
         $errorMsg = "Not found";
         return view('admin.error', compact('errorMsg'));
@@ -65,7 +70,15 @@ class ProductController extends Controller
 
     public function delete(Request $request) {
         if($request->input('id')) {
-            Product::destroy($request->input('id'));
+            // check if product have order. do not allow
+            $orderItem = OrderItem::where('product_id', $request->input('id'))->first();
+            if($orderItem) {
+                $errorMsg = "The product that your try to delete have order record. Delete will break order report. Instead of delete, DISABLE the product. ";
+                return view('admin.error', compact('errorMsg'));
+            }
+            else {
+                Product::destroy($request->input('id'));
+            }
         }
         return view('admin.productlist');
     }
